requirements.txt
dnspython
flask
gunicorn[eventlet]
django
requests
mysql-connector-python
nslookup